+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added parameters to verified app link
 - Renamed hard and soft flow to login and permission
 - Implemented new design
+- Claims are now a string
 
 ### Fixed
 - Fixed UserInfo json parser error